personal/knowledge-engine/pyproject.toml
[project]
name = "knowledge-engine"
version = "0.1.0"
description = "Enterprise-grade knowledge infrastructure for AI applications"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
keywords = ["knowledge", "rag", "graphrag", "ollama", "mcp", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# ============================================================
# CORE DEPENDENCIES (FREE tier)
# ============================================================
dependencies = [
    # API Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",

    # Database Clients (self-hosted = FREE)
    "qdrant-client>=1.12.0",
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",

    # HTTP Client (for Ollama)
    "httpx>=0.28.0",

    # Utilities
    "tenacity>=9.0.0",
    "python-multipart>=0.0.17",
    "structlog>=24.4.0",

    # Content Ingestion
    "trafilatura>=1.12.0",
    "youtube-transcript-api>=0.6.3",
    "pypdf>=5.1.0",

    # Metrics (used in middleware)
    "prometheus-client>=0.21.0",
]

[project.optional-dependencies]
# ============================================================
# OPTIONAL: Graph Database (Neo4j)
# Install: pip install knowledge-engine[graph]
# ============================================================
graph = [
    "neo4j>=5.26.0",
]

# ============================================================
# OPTIONAL: Premium AI Providers
# Install: pip install knowledge-engine[premium]
# ============================================================
premium = [
    "voyageai>=0.3.0",
    "cohere>=5.11.0",
    "anthropic>=0.40.0",
]

# ============================================================
# OPTIONAL: MCP Server
# Install: pip install knowledge-engine[mcp]
# ============================================================
mcp = [
    "mcp>=1.2.0",
]

# ============================================================
# OPTIONAL: Redis Cache
# Install: pip install knowledge-engine[cache]
# ============================================================
cache = [
    "redis>=5.0.0",
]

# ============================================================
# OPTIONAL: Observability
# Install: pip install knowledge-engine[observability]
# ============================================================
observability = [
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-instrumentation-httpx>=0.49b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.49b0",
    "opentelemetry-instrumentation-redis>=0.49b0",
    "prometheus-client>=0.21.0",
]

# ============================================================
# ALL OPTIONAL FEATURES
# Install: pip install knowledge-engine[all]
# ============================================================
all = [
    "knowledge-engine[graph]",
    "knowledge-engine[premium]",
    "knowledge-engine[mcp]",
    "knowledge-engine[cache]",
    "knowledge-engine[observability]",
]

# ============================================================
# DEVELOPMENT
# Install: pip install knowledge-engine[dev]
# ============================================================
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
    "respx>=0.22.0",
]

[project.scripts]
knowledge-engine-mcp = "knowledge_engine.mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/knowledge_engine"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # Line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["knowledge_engine"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "voyageai.*",
    "cohere.*",
    "neo4j.*",
    "qdrant_client.*",
    "mcp.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
branch = true
